Provide a stub ethernet service to build in sc-mainline-prod.
This CL also creates a fake ethernet-service.jar to ensure that
targets that depend on ethernet-service do not break.
Because ethernet-service.jar is on the system server classpath,
these classes cannot be in the com.android.server.ethernet
package, otherwise ConnectivityServiceInitializer will load
them instead of the real classes.

Test: m
Bug: 210586283
Ignore-AOSP-First: avoid merge conflicts
Change-Id: I6a4a0dc68233d4229d1cae4e5b196e287ef2ed5b