Deflake testStartStopVpnProfileV6 and testStartStopVpnProfileV4
The tests expect to receive network validated callback on VPN
network in S and below. However, there is a platform bug that
VPN may misuse VPN network itself as the underlying network,
and cause VPN reconnect. If this race happens before validation
completed and callback sent, the test will fail to get the
validated network callback.

Restrict the validated capabilities check in T+ only devices
which contain the necessary fixes.

Also, unreigster the callback at the end of the test.

Bug: 228790797
Test: atest android.net.cts.Ikev2VpnTest
Change-Id: Id0ae0f983f989bfde5df2481e0d529bc5a140a7d