Improve count of openSession callers in dumpsys
Currently, it's only uid=count in dumpsys, however, the call may also
come from system but there is no such information in dumpsys. If the
call came from system, it's hard to know who is the caller. So, this
patch makes pid, uid and package as a key to record the count of the
caller.

The output will look like,
Top openSession callers:
  {pid=1787,uid=1000,package=android}=26
  {pid=2931,uid=1000,package=com.android.settings}=4

Test: adb shell dumpsys netstats and check the output
Bug: 228081549
Change-Id: Id2dfdd4480aa5b1ece51e46de6efe30a46629811