Enable tests for firewall chain blocking
Tests were introduced in aosp/2114533. But we disabled them because they
affect the actual state of networking on the device after the tests and
this might cause unexpected behavior of the device.

This commit updates tests to restore the status after the tests and
enables tests.

Bug: 208371987
Test: atest
CtsNetTestCases: android.net.cts.ConnectivityManagerTest#testFirewallBlocking
Change-Id: I7e3ebdff7d2d8a227291ac12c4fe72529986e74d