ethernet: rename mDefaultInterface to mTetheringInterface
The name "default interface" is misleading as "default" is an overloaded
term that can imply many things. Renaming it to mTetheringInterface, so
it becomes clearer what this interface is used for. I briefly considered
naming it mInterfaceUsedForTethering, but then
mInterfaceUsedForTetheringMode becomes annoyingly long. Decided to go
with mTetheringInterface and mTetheringInterfaceMode instead. This
change also gets rid of all mentions of default in the context of
tethering.

Test: m
Bug: 236280707
Change-Id: Ia5ddbdc5a8b258d7fb050126288b82a1c8cf666d