Update xTS owners for cherry-pick purposes
Add owners to approve xTS changes that are cherry-picks
of changes already merged in aosp/master, as suggested
by current owners.

Also add LAST_RESORT_SUGGESTION to everyone except
lorenzo@, as only lorenzo@ is expected to approve the
original changes, so suggesting other reviewers would
most of the time not be helpful.

Change-Id: I3d04eab6dc5070e09840b9b9563d8955ed18669d
Test: m