Increase timeout for waiting a validated network
testDefaultNetworkActiveListener expects Wifi network to be
the system default network. This requires Wifi to be validated
since the cellular network may be connected in the devices that
support telephony. The original timeout was only set to 5
seconds. The timeout may not be enough in some poor environment
cases, so increase the timeout would be helpful to reduce the
flaky rate.

Bug: 250250617
Test: atest android.net.cts.ConnectivityManagerTest
Change-Id: Iaef307e72dcd7a74cd5fdc6a406443eed6b412a6