Make mutable implicit PendingIntents explicit and immutable
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the the first PendingIntent was
destined to the test app/to the app, so it was made explicit. The second
PendingIntent is not mutated here, so was made immutable.

Reviewers, please call out if this is not the case.

Bug: 236704164
Bug: 229362273
Test: atest NearbyUnitTests
Ignore-AOSP-First: nearby_not_in_aosp_yet
Change-Id: I78126425d1da65fc65ce36ae80c6d72988c632ff