cronet import: use folder.origin instead of git.origin
git.origin uses a bare repo which is not supported by gclient. There may
be ways to get this to work including the git-origin-checkout-hook, but
those are a) really slow and b) much more complex than just using a
folder.origin.

This change also adds support to the import script to configure the
origin folder (i.e. clone the repo and run gclient) before the copybara
script is invoked.

Test: run import script
Change-Id: I6289108271960c87583a059a553b0d0e7db8af68