Move fail in testFirewallBlocking
testFirewallBlocking is flaky and failure was that pkt is expected to be
dropped but pkt was not dropped.
Root cause of this failure was not found out and possible next actions
were discussed in b/276496378.

Currently, test expect DatagramSocket.send throws exception when pkt is
expected to be dropped and test fail just after send does not throw.
This CL moves fail after pkt is received for future investigation.

Bug: 276496378
Test: atest ConnectivityManagerTest
Change-Id: Ife8b0da13843e56b9d5465d8d0464d0ae396246c